Load LOINC Table File (CSV) into Postgres

The CSV files unpacked from a LOINC Table File (CSV) download from https://loinc.org/downloads/loinc-table/ are loaded into Postgres.The file version and release dates are taken along with the table names and row counts are logged to a separate table after the loading is completed.

run_setup(
  conn,
  schema = "loinc",
  csv_path,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  log_schema = "public",
  log_table_name = "setup_loinc_log",
  log_csv_version,
  log_release_date
)

Arguments

conn
Connection to a Postgres database.

schema
Target schema for the LOINC load, Default: 'loinc'

csv_path
Path to the unpacked LOINC Table File csv files.

log_schema
Schema for the table that logs the process, Default: 'public'

log_table_name
Name of log table, Default: 'setup_loinc_log'

log_csv_version
(Required) https://loinc.org/downloads/loinc-table/ Version.

log_release_date
(Required) https://loinc.org/downloads/loinc-table/ Release Date.

See also